rygg/requirements.txt
asgiref>=3.3.2
blessings>=1.7
celery>=5.1.0
certifi>=2020.6.20
chardet>=3.0.4
curtsies>=0.3.1
Deprecated>=1.2.10
django-cors-headers>=3.2.1
django-extensions>=3.1.3
django-http-exceptions>=1.3.1
django-model-utils>=4.1.1
django-request-logging>0.7.2
Django==3.2
djangorestframework>=3.11.1
flower>=1.0.0
gitdb>=4.0.5
GitPython>=3.1.8
greenlet>=0.4.15
idna>=2.10
importlib-metadata>=1.7.0
markdown>=3.2.1
more-itertools>=8.9.0
oauthlib>=3.0.2
packaging>=20.4
psycopg2-binary>=2.8.5
PyGithub>=1.53
pygments>=2.6.1
PyJWT>=1.7.1
pyparsing>=2.4.7
pypiwin32>=223; platform_system == "Windows"
python-dateutil>=2.8.1
pytz>=2020.1
redis>=3.5.3
requests-oauthlib>=1.2.0
requests>=2.24.0
retrying>=1.3.3
Send2Trash>=1.5.0
six>=1.14.0
smmap>=3.0.4
sqlparse>=0.3.1
urllib3==1.25.10 #pinning this to 1.25.10 to avoid installation errors for the packages 'requests' and 'botocore'
wcwidth>=0.1.9
zipp>=3.1.0
